refactor(config): move settings from data_processor.py to config.py
